Add Single File Format (SFF) to ease testing (#7)
Add a simple file format that stores all of the input data in a
single file. This allows either simpler testing with the GOTCHA
data (by converting a set of files into a single file) or
testing with other data. The reader and writer also support a
mode that only writes metadata and generates random data on
input. This is useful for benchmarking using small input files.

Signed-off-by: Thomas Benson <tbensongit@gmail.com>